scheduler/requirements.txt
kafka-python==2.0.2
apscheduler==3.10.4